# Conflicts:
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/AutofillStrikeDatabase/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/AutofillStrikeDatabase/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/BudgetDatabase/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/BudgetDatabase/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Cache/data_0
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Cache/data_1
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Cache/data_2
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Cache/data_3
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Code Cache/js/index-dir/the-real-index
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Extension State/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Extension State/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Favicons
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Favicons-journal
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Feature Engagement Tracker/AvailabilityDB/000003.log
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Feature Engagement Tracker/AvailabilityDB/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Feature Engagement Tracker/AvailabilityDB/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Feature Engagement Tracker/EventDB/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Feature Engagement Tracker/EventDB/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/History
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/History-journal
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Login Data
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Preferences
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Site Characteristics Database/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Site Characteristics Database/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Sync Data/LevelDB/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Sync Data/LevelDB/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/Visited Links
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/data_reduction_proxy_leveldb/CURRENT
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/data_reduction_proxy_leveldb/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/data_reduction_proxy_leveldb/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/shared_proto_db/metadata/LOG
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Default/shared_proto_db/metadata/LOG.old
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/Local State
#	eye-tracking-for-simulation-assessment/simulation/chromium/userData/ShaderCache/GPUCache/data_1

+++ eye-tracking-for-simulation-assessment/src/background.js
@@ -285,6 +285,12 @@ function readMapFile(event, file) {
     filePath = path.join('./resources/simulation/', file + '.csv')
   }
 
+  if (!fs.existsSync(filePath)) {
+    console.log('csv文件不存在')
+    event.reply('mapLoaded' + file, [])
+    return
+  }
+
   fs.readFile(filePath, (err, data) => {
     if (err) {
       console.log(err.stack)
@@ -313,11 +319,16 @@ function renameFile(event, name) {
   } else {
     dir = path.resolve('./resources/simulation/')
   }
+
   let oldPath = path.join(dir, 'std.csv')
   let newPath = path.join(dir, name + 'STD' + '.csv')
 
-  console.log(oldPath)
-  console.log(newPath)
+  if (!fs.existsSync(oldPath)) {
+    console.log('csv文件不存在')
+    event.reply('renamed' + name)
+    return
+  }
+
   fs.rename(oldPath, newPath, (err) => {
     if (err) {
       console.log(err.stack)
@@ -327,6 +338,13 @@ function renameFile(event, name) {
 
   oldPath = path.join(dir, 'count.csv')
   newPath = path.join(dir, name + '.csv')
+
+  if (!fs.existsSync(oldPath)) {
+    console.log('csv文件不存在')
+    event.reply('renamed' + name)
+    return
+  }
+
   fs.rename(oldPath, newPath, (err) => {
     if (err) {
       console.log(err.stack)
